Beta 1.0.20/1.1.0(iOS)

+++ lib/views/my_page/setting/setting_page.dart
@@ -1,9 +1,9 @@
+import 'dart:io';
+
 import 'package:flutter/material.dart';
-import 'package:flutter_webview_plugin/flutter_webview_plugin.dart';
-import 'package:scoped_model/scoped_model.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:webview_flutter/webview_flutter.dart';
-import 'package:yuyan_app/models/component/appUI.dart';
+import 'package:yuyan_app/models/widgets_small/show_dialog/show_confirm.dart';
 import 'package:yuyan_app/state_manage/toppest.dart';
 import 'package:yuyan_app/views/begin_init/login_page/login_page.dart';
 import 'package:yuyan_app/views/my_page/setting/view/one_tile.dart';
@@ -14,6 +14,7 @@ class SettingPage extends StatefulWidget {
 
 class _SettingPageState extends State<SettingPage> {
   checkVersion() {
+    if ((Platform.isIOS)) {}
     topModel.versionManage.checkVersion(context);
   }
 
@@ -43,15 +44,18 @@ class _SettingPageState extends State<SettingPage> {
             title: '退出登录',
             icon: Icons.power_settings_new,
             onTap: () async {
-              SharedPreferences prefs = await SharedPreferences.getInstance();
-              prefs.clear();
-              CookieManager cookieManager = CookieManager();
-              await cookieManager.clearCookies();
-              Navigator.of(context).pushAndRemoveUntil(
-                  new MaterialPageRoute(
-                    builder: (BuildContext context) => new LoginPage(),
-                  ),
-                  (Route route) => route == null);
+              showConfirmDialog(context, content: "确认要退出登录 ?",
+                  confirmCallback: () async {
+                SharedPreferences prefs = await SharedPreferences.getInstance();
+                prefs.clear();
+                CookieManager cookieManager = CookieManager();
+                await cookieManager.clearCookies();
+                Navigator.of(context).pushAndRemoveUntil(
+                    new MaterialPageRoute(
+                      builder: (BuildContext context) => new LoginPage(),
+                    ),
+                    (Route route) => route == null);
+              });
             },
           ),
         ],
